Add support for visual mode selection.
The regex in ghcimportedfrom#getHaskellIdentifier() doesn't work
if we have the cursor over a weird symbol like

    <$

so alternatively the user may use visual mode to select the text of the
symbol and then call one of the alternatively named functions ending in
Vismode.

We probably can't extend the regex to include ```<``` and so on as this
would make '<x' match to the whole '<x' term with the cursor over x, but
it should only match the 'x'.